Fix issue where graphs couldn't be fetched anymore.
The problem was fixed by calling the API twice in a row when we get redirected.
Apparently, the 302 just wants us to... keep away for a while before we retry?
Don't even need to follow the redirection. Just retry. That's nice.

Also keep all cookies in a session because that helps.